NMADLC

+++ src/controllers/Supplier.controller.js
@@ -1,3 +1,4 @@
+import { shopping } from '../models/shopping.model.js';
 import {supplier} from '../models/supplier.model.js'
 
 export const getSupplier = async (req, res) => {
@@ -89,6 +90,18 @@ export const deleteSupplier = async (req, res) => {
     try {
         const { id } = req.params;
         
+        const existSupplirInShoppings = await shopping.findOne({
+            where: {
+                Supplier_ID: id
+            }
+        })
+
+        if (existSupplirInShoppings) {
+            return res.status(403).json({
+                message: "El provedor no puede ser eliminado",
+                useDelete: false
+            })
+        }
         await supplier.destroy({
             where: {
                 ID_Supplier: id
